requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

coloredlogs==15.0.1
ijson==3.5.0
numpy==2.4.0
pillow==12.1.0
requests==2.32.5
rich==14.2.0
terminaltables3==4.0.0
